VERY WIP: going to rewrite how tags are represented and manipulated

+++ src/functions/neighbourhood.rs
@@ -2,11 +2,11 @@
 ///
 
 use ndarray;
-use ndarray::{Array2};
+use ndarray::Array2;
 use std::f64::consts::PI as PI;
 
 // Default neighbourhood function: Gaussian function; returns a Gaussian centered in pos
-pub fn gaussian(size: (usize, usize), pos: (usize, usize), sigma: f32) -> Array2<f64> {
+pub fn gaussian(size: (usize, usize), pos: (usize, usize), sigma: f64) -> Array2<f64> {
     let mut ret = Array2::<f64>::zeros((size.0, size.1));
     let div = 2.0 * PI * sigma as f64 * sigma as f64;
